Separating the headers so that the inline functions are not messing up the declarations.
