+++ include/ada.h
@@ -1,18 +1,18 @@
 #ifndef ADA_H
 #define ADA_H
 
-#include "ada/character_sets.h"
-#include "ada/checkers.h"
+#include "ada/character_sets-inl.h"
+#include "ada/checkers-inl.h"
 #include "ada/common_defs.h"
 #include "ada/log.h"
 #include "ada/encoding_type.h"
 #include "ada/helpers.h"
 #include "ada/parser.h"
-#include "ada/scheme.h"
+#include "ada/scheme-inl.h"
 #include "ada/serializers.h"
 #include "ada/state.h"
 #include "ada/unicode.h"
-#include "ada/url.h"
+#include "ada/url-inl.h"
 
 // Public API
 #include "ada/ada_version.h"
