+++ include/ada/checkers.h
@@ -1,6 +1,6 @@
 /**
  * @file checkers.h
- * @brief Definitions for URL specific checkers used within Ada.
+ * @brief Declarations for URL specific checkers used within Ada.
  */
 #ifndef ADA_CHECKERS_H
 #define ADA_CHECKERS_H
@@ -12,13 +12,11 @@
 
 namespace ada::checkers {
 
-  //
-  // More likely to be inlined by the compiler and constexpr.
   /**
    * Assuming that x is an ASCII letter, this function returns the lower case equivalent.
    * @details More likely to be inlined by the compiler and constexpr.
    */
-  constexpr char to_lower(char x) noexcept { return (x | 0x20); }
+  constexpr char to_lower(char x) noexcept;
 
   /**
    * Returns true if the character is an ASCII letter. Equivalent to std::isalpha but
@@ -26,7 +24,7 @@ namespace ada::checkers {
    *
    * @attention std::isalpha is not constexpr generally.
    */
-  constexpr bool is_alpha(char x) noexcept { return (to_lower(x) >= 'a') && (to_lower(x) <= 'z'); }
+  constexpr bool is_alpha(char x) noexcept;
 
   /**
    * Check whether a string starts with 0x or 0X. The function is only
@@ -34,29 +32,16 @@ namespace ada::checkers {
    *
    * @see has_hex_prefix
    */
-  inline bool has_hex_prefix_unsafe(std::string_view input) {
-    // This is actualy efficient code, see has_hex_prefix for the assembly.
-    uint32_t value_one = 1;
-    bool is_little_endian = (reinterpret_cast<char*>(&value_one)[0] == 1);
-    uint16_t word0x{};
-    std::memcpy(&word0x, "0x", 2); // we would use bit_cast in C++20 and the function could be constexpr.
-    uint16_t two_first_bytes{};
-    std::memcpy(&two_first_bytes, input.data(),2);
-    if(is_little_endian) { two_first_bytes |= 0x2000; } else { two_first_bytes |= 0x020; }
-    return two_first_bytes == word0x;
-  }
-
+  inline bool has_hex_prefix_unsafe(std::string_view input);
   /**
    * Check whether a string starts with 0x or 0X.
    */
-  inline bool has_hex_prefix(std::string_view input) {
-    return input.size() >=2 && has_hex_prefix_unsafe(input);
-  }
+  inline bool has_hex_prefix(std::string_view input);
 
   /**
    * Check whether x is an ASCII digit. More likely to be inlined than std::isdigit.
    */
-  constexpr bool is_digit(char x) noexcept { return (x >= '0') & (x <= '9'); }
+  constexpr bool is_digit(char x) noexcept;
 
   /**
    * @details A string starts with a Windows drive letter if all of the following are true:
@@ -67,25 +52,17 @@ namespace ada::checkers {
    *
    * https://url.spec.whatwg.org/#start-with-a-windows-drive-letter
    */
-  inline constexpr bool is_windows_drive_letter(std::string_view input) noexcept {
-    return input.size() >= 2 && (is_alpha(input[0]) && ((input[1] == ':') || (input[1] == '|')))
-      && ((input.size() == 2) || (input[2] == '/' || input[2] == '\\' || input[2] == '?' || input[2] == '#'));
-  }
+  inline constexpr bool is_windows_drive_letter(std::string_view input) noexcept;
 
   /**
    * @details A normalized Windows drive letter is a Windows drive letter of which the second code point is U+003A (:).
    */
-  inline constexpr bool is_normalized_windows_drive_letter(std::string_view input) noexcept {
-    return input.size() >= 2 && (is_alpha(input[0]) && (input[1] == ':'));
-  }
+  inline constexpr bool is_normalized_windows_drive_letter(std::string_view input) noexcept;
 
   /**
    * @warning Will be removed when Ada supports C++20.
    */
-  ada_really_inline constexpr bool begins_with(std::string_view view, std::string_view prefix) {
-    // in C++20, you have view.begins_with(prefix)
-    return view.size() >= prefix.size() && (view.substr(0, prefix.size()) == prefix);
-  }
+  ada_really_inline constexpr bool begins_with(std::string_view view, std::string_view prefix);
 
   /**
    * Returns true if an input is an ipv4 address.
