+++ include/ada/scheme.h
@@ -1,6 +1,6 @@
 /**
  * @file scheme.h
- * @brief Definitions for the URL scheme.
+ * @brief Declarations for the URL scheme.
  */
 #ifndef ADA_SCHEME_H
 #define ADA_SCHEME_H
@@ -28,15 +28,6 @@ namespace ada::scheme {
     FILE = 6
   };
 
-  namespace details {
-    // for use with is_special and get_special_port
-    // Spaces, if present, are removed from URL.
-    constexpr std::string_view is_special_list[] = {"http", " ", "https",
-                                                    "ws", "ftp", "wss", "file", " "};
-    // for use with get_special_port
-    constexpr uint16_t special_ports[] = {80, 0, 443, 80, 21, 443, 0, 0};
-  }
-
   /**
    * A special scheme is an ASCII string that is listed in the first column of the following table.
    * The default port for a special scheme is listed in the second column on the same row.
@@ -46,12 +37,7 @@ namespace ada::scheme {
    * @param scheme
    * @return If scheme is a special scheme
    */
-  ada_really_inline constexpr bool is_special(std::string_view scheme) {
-    if(scheme.empty()) { return false; }
-    int hash_value = (2*scheme.size() + (unsigned)(scheme[0])) & 7;
-    const std::string_view target = details::is_special_list[hash_value];
-    return (target[0] == scheme[0]) && (target.substr(1) == scheme.substr(1));
-  }
+  ada_really_inline constexpr bool is_special(std::string_view scheme);
 
   /**
    * A special scheme is an ASCII string that is listed in the first column of the following table.
@@ -62,34 +48,17 @@ namespace ada::scheme {
    * @param scheme
    * @return The special port
    */
-  constexpr uint16_t get_special_port(std::string_view scheme) noexcept {
-    if(scheme.empty()) { return 0; }
-    int hash_value = (2*scheme.size() + (unsigned)(scheme[0])) & 7;
-    const std::string_view target = details::is_special_list[hash_value];
-    if ((target[0] == scheme[0]) && (target.substr(1) == scheme.substr(1))) {
-      return details::special_ports[hash_value];
-    } else { return 0; }
-  }
+  constexpr uint16_t get_special_port(std::string_view scheme) noexcept;
 
   /**
    * Returns the port number of a special scheme.
    * @see https://url.spec.whatwg.org/#special-scheme
    */
-  constexpr uint16_t get_special_port(ada::scheme::type type) noexcept {
-    return details::special_ports[int(type)];
-  }
-
+  constexpr uint16_t get_special_port(ada::scheme::type type) noexcept;
   /**
    * Returns the scheme of an input, or NOT_SPECIAL if it's not a special scheme defined by the spec.
    */
-  constexpr ada::scheme::type get_scheme_type(std::string_view scheme) noexcept {
-    if(scheme.empty()) { return ada::scheme::NOT_SPECIAL; }
-    int hash_value = (2*scheme.size() + (unsigned)(scheme[0])) & 7;
-    const std::string_view target = details::is_special_list[hash_value];
-    if ((target[0] == scheme[0]) && (target.substr(1) == scheme.substr(1))) {
-      return ada::scheme::type(hash_value);
-    } else { return ada::scheme::NOT_SPECIAL; }
-  }
+  constexpr ada::scheme::type get_scheme_type(std::string_view scheme) noexcept;
 
 } // namespace ada::serializers
