+++ include/ada/url.h
@@ -1,6 +1,6 @@
 /**
  * @file url.h
- * @brief Definitions for the URL
+ * @brief Declaration for the URL
  */
 #ifndef ADA_URL_H
 #define ADA_URL_H
@@ -27,6 +27,13 @@ namespace ada {
    * @see https://url.spec.whatwg.org/#url-representation
    */
   struct url {
+    url() = default;
+    url(const url &u) = default;
+    url(url &&u) noexcept = default;
+    url &operator=(url &&u) noexcept = default;
+    url &operator=(const url &u) = default;
+    ADA_ATTRIBUTE_NOINLINE ~url() = default;
+
     /**
      * @private
      * A URL’s username is an ASCII string identifying a username. It is initially the empty string.
@@ -215,16 +222,12 @@ namespace ada {
     /**
      * A URL includes credentials if its username or password is not the empty string.
      */
-    [[nodiscard]] ada_really_inline bool includes_credentials() const noexcept {
-      return !username.empty() || !password.empty();
-    }
+    [[nodiscard]] ada_really_inline bool includes_credentials() const noexcept;
 
     /**
      * A URL is special if its scheme is a special scheme. A URL is not special if its scheme is not a special scheme.
      */
-    [[nodiscard]] ada_really_inline bool is_special() const noexcept {
-      return type != ada::scheme::NOT_SPECIAL;
-    }
+    [[nodiscard]] ada_really_inline bool is_special() const noexcept;
 
     /**
      * @private
@@ -232,9 +235,7 @@ namespace ada {
      * Return the 'special port' if the URL is special and not 'file'.
      * Returns 0 otherwise.
      */
-    [[nodiscard]] uint16_t get_special_port() const {
-      return ada::scheme::get_special_port(type);
-    }
+    [[nodiscard]] inline uint16_t get_special_port() const;
 
     /**
      * @private
@@ -244,34 +245,20 @@ namespace ada {
      * get_scheme() == "file", since the former is a direct integer comparison,
      * while the other involves a (cheap) string test.
      */
-    [[nodiscard]] ada_really_inline ada::scheme::type get_scheme_type() const noexcept {
-      return type;
-    }
+    [[nodiscard]] ada_really_inline ada::scheme::type get_scheme_type() const noexcept;
 
     /**
      * @private
      *
      * Get the default port if the url's scheme has one, returns 0 otherwise.
      */
-    [[nodiscard]] ada_really_inline uint16_t scheme_default_port() const noexcept {
-      return scheme::get_special_port(type);
-    }
-
+    [[nodiscard]] ada_really_inline uint16_t scheme_default_port() const noexcept;
     /**
      * @private
      *
      * A URL cannot have a username/password/port if its host is null or the empty string, or its scheme is "file".
      */
-    [[nodiscard]] bool cannot_have_credentials_or_port() const {
-      return !host.has_value() || host.value().empty() || type == ada::scheme::type::FILE;
-    }
-    /** For development purposes, we want to know when a copy is made. */
-    url() = default;
-    url(const url &u) = default;
-    url(url &&u) noexcept = default;
-    url &operator=(url &&u) noexcept = default;
-    url &operator=(const url &u) = default;
-    ADA_ATTRIBUTE_NOINLINE ~url() = default;
+    [[nodiscard]] inline bool cannot_have_credentials_or_port() const;
 
     /**
      * @private
@@ -283,28 +270,7 @@ namespace ada {
      * @return On failure, it returns zero.
      * @see https://url.spec.whatwg.org/#host-parsing
      */
-    ada_really_inline size_t parse_port(std::string_view view, bool check_trailing_content = false) noexcept {
-      ada_log("parse_port('", view, "') ", view.size());
-      uint16_t parsed_port{};
-      auto r = std::from_chars(view.data(), view.data() + view.size(), parsed_port);
-      if(r.ec == std::errc::result_out_of_range) {
-        ada_log("parse_port: std::errc::result_out_of_range");
-        is_valid = false;
-        return 0;
-      }
-      ada_log("parse_port: ", parsed_port);
-      const size_t consumed = size_t(r.ptr - view.data());
-      ada_log("parse_port: consumed ", consumed);
-      if(check_trailing_content) {
-        is_valid &= (consumed == view.size() || view[consumed] == '/' || view[consumed] == '?' || (is_special() && view[consumed] == '\\'));
-      }
-      ada_log("parse_port: is_valid = ", is_valid);
-      if(is_valid) {
-        port = (r.ec == std::errc() && scheme_default_port() != parsed_port) ?
-          std::optional<uint16_t>(parsed_port) : std::nullopt;
-      }
-      return consumed;
-    }
+    ada_really_inline size_t parse_port(std::string_view view, bool check_trailing_content = false) noexcept;
 
     /**
      * @private
@@ -312,24 +278,13 @@ namespace ada {
      * Return a string representing the scheme. Note that get_scheme_type() should often be used instead.
      * @see https://url.spec.whatwg.org/#dom-url-protocol
      */
-    [[nodiscard]] std::string_view get_scheme() const noexcept {
-      if(is_special()) { return ada::scheme::details::is_special_list[type]; }
-      // We only move the 'scheme' if it is non-special.
-      return non_special_scheme;
-    }
-
+    [[nodiscard]] inline std::string_view get_scheme() const noexcept;
     /**
      * Set the scheme for this URL. The provided scheme should be a valid
      * scheme string, be lower-cased, not contain spaces or tabs. It should
      * have no spurious trailing or leading content.
      */
-    void set_scheme(std::string&& new_scheme) noexcept {
-      type = ada::scheme::get_scheme_type(new_scheme);
-      // We only move the 'scheme' if it is non-special.
-      if(!is_special()) {
-        non_special_scheme = new_scheme;
-      }
-    }
+    inline void set_scheme(std::string&& new_scheme) noexcept;
 
     /**
      * @private
@@ -337,10 +292,7 @@ namespace ada {
      * Take the scheme from another URL. The scheme string is moved from the
      * provided url.
      */
-    void copy_scheme(ada::url&& u) noexcept {
-      non_special_scheme = u.non_special_scheme;
-      type = u.type;
-    }
+    inline void copy_scheme(ada::url&& u) noexcept;
 
     /**
      * @private
@@ -348,10 +300,7 @@ namespace ada {
      * Take the scheme from another URL. The scheme string is copied from the
      * provided url.
      */
-    void copy_scheme(const ada::url& u) {
-      non_special_scheme = u.non_special_scheme;
-      type = u.type;
-    }
+    inline void copy_scheme(const ada::url& u);
 
     /**
      * @private
@@ -451,9 +400,7 @@ namespace ada {
   }; // struct url
 
 
-  inline std::ostream& operator<<(std::ostream& out, const ada::url& u) {
-    return out << u.to_string();
-  }
+  inline std::ostream& operator<<(std::ostream& out, const ada::url& u);
 } // namespace ada
 
 #endif // ADA_URL_H
