+++ src/checkers.cpp
@@ -1,4 +1,4 @@
-#include "ada.h"
+#include "ada/checkers.h"
 #include <algorithm>
 
 namespace ada::checkers {
